LayoutTests/imported/mozilla/css-transitions/test_animation-pausing-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: The operation was aborted.
#PID UNRESPONSIVE - com.apple.WebKit.WebContent.Development (pid 13519)
FAIL: Timed out waiting for notifyDone to be called
